+++ src/main/java/rsystems/Handlers/SQLHandler.java
@@ -2,6 +2,8 @@ package rsystems.handlers;
 
 import java.sql.*;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 public class SQLHandler {
 
@@ -17,6 +19,7 @@ public class SQLHandler {
         }
     }
 
+    // Get date of last seen using ID
     public String getDate(String id){
         String date = "";
 
@@ -42,6 +45,7 @@ public class SQLHandler {
         return date;
     }
 
+    // Set the last seen date using id
     public int setDate(String id, String date){
         try{
 
@@ -60,6 +64,33 @@ public class SQLHandler {
         return 0;
     }
 
+    // Get date of last seen using ID
+    public String getName(String id){
+        String name = "";
+
+        try{
+
+            if((connection == null) || (connection.isClosed())){
+                connect();
+            }
+
+            Statement st = connection.createStatement();
+
+            ResultSet rs = st.executeQuery("SELECT NAME FROM LastSeenTable WHERE ID = " + id);
+            while(rs.next()){
+                name = rs.getString("NAME");
+            }
+
+        } catch (SQLException throwables) {
+            System.out.println(throwables.getMessage());
+        } finally {
+            closeConnection();
+        }
+
+        return name;
+    }
+
+    // Set the name of the user using id
     public int setName(String id, String name){
         try{
 
@@ -78,6 +109,7 @@ public class SQLHandler {
         return 0;
     }
 
+    // Insert NEW users into the DB
     public boolean insertUser(String id, String name, String date){
         try{
 
@@ -97,6 +129,7 @@ public class SQLHandler {
         return false;
     }
 
+    // Remove using from DB
     public boolean removeUser(String id){
         try{
 
@@ -116,6 +149,7 @@ public class SQLHandler {
         return false;
     }
 
+    // Close Connection to DB
     private void closeConnection(){
         try {
             if (connection != null) {
@@ -126,8 +160,9 @@ public class SQLHandler {
         }
     }
 
+    // Return an array list of all values in the DB
     public ArrayList<String> getAllUsers(String column) {
-        ArrayList<String> users = new ArrayList<>();
+        ArrayList<String> values = new ArrayList<>();
 
         try {
 
@@ -139,7 +174,7 @@ public class SQLHandler {
 
             ResultSet rs = st.executeQuery("SELECT " + column + " FROM LastSeenTable");
             while (rs.next()) {
-                users.add(rs.getString(column));
+                values.add(rs.getString(column));
             }
 
         } catch (SQLException throwables) {
@@ -148,7 +183,56 @@ public class SQLHandler {
             closeConnection();
         }
 
-        return users;
+        return values;
+    }
+
+    // Return an array list of all values in the DB
+    public HashMap<String,String> getAllUsers() {
+        HashMap<String,String> idMap = new HashMap<>();
+
+        try {
+
+            if ((connection == null) || (connection.isClosed())) {
+                connect();
+            }
+
+            Statement st = connection.createStatement();
+
+            ResultSet rs = st.executeQuery("SELECT ID, NAME FROM LastSeenTable");
+            while (rs.next()) {
+                idMap.put(rs.getString("ID"),rs.getString("NAME"));
+            }
+
+        } catch (SQLException throwables) {
+            System.out.println(throwables.getMessage());
+        } finally {
+            closeConnection();
+        }
+
+        return idMap;
+    }
+
+    public ResultSet getAllData(){
+        ResultSet rs = null;
+
+        try {
+
+            if ((connection == null) || (connection.isClosed())) {
+                connect();
+            }
+
+            Statement st = connection.createStatement();
+
+            rs = st.executeQuery("SELECT * FROM LastSeenTable");
+            //return rs;
+
+        } catch (SQLException throwables) {
+            System.out.println(throwables.getMessage());
+        } finally {
+            closeConnection();
+        }
+
+        return rs;
     }
 
 }